pyproject.toml
[project]
name = "resume"
version = "0.1.0"
description = "Professional resume website built with Streamlit"
authors = [
    {name = "chriztopherton",email = "christopher.ton@sjsu.edu"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "streamlit-option-menu>=0.3.6",
    "streamlit-extras>=0.3.6",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain (>=0.3.27,<0.4.0)"
]

[project.optional-dependencies]
dev = [
    "commitizen>=4.8.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
version_files = [
    "pyproject.toml:version"
]
